refactor: spliting into modular monolith
