+++ .env.test
@@ -2,7 +2,7 @@
 TEST_FILE_PATH = "../../data/emails_test.txt"
 EMPTY_TEST_FILE_PATH = "../../data/empty_emails_test.txt"
 FULL_TEST_FILE_PATH = "../../data/full_emails_test.txt"
-REPO_TEST_FILE_PATH = "../../data/subs_repo_test.txt"
+REPO_TEST_FILE_PATH = "../../../data/subs_repo_test.txt"
 EMAILS_FILE_PATH = "data/emails.txt"
 
 # Currencies
@@ -10,12 +10,12 @@ BASE_CURRENCY = "BTC"
 QUOTE_CURRENCY = "UAH"
 
 # SMTP server
-SENDER_EMAIL = "YOUR_VALUE"
-SENDER_PASSWORD = "YOUR_VALUE"
+SENDER_EMAIL = "artem.kapustkin.work@gmail.com"
+SENDER_PASSWORD = "jyvyifnqdsxzhgin"
 
 # Providers
 BINANCE_BASE_URL = "https://api.binance.com/api/v3/ticker"
 COIN_API_BASE_URL = "https://rest.coinapi.io/v1/exchangerate"
 COINBASE_BASE_URL = "https://api.coinbase.com/v2"
-COIN_API_KEY = "YOUR_VALUE"
+COIN_API_KEY = "DADD6924-3A3B-40E0-B7A2-61B8815A46F8"
 DEFAULT_PROVIDER_NAME = "binance"
