+++ modules/notification_module/model/email_message.go
