+++ modules/rate_module/model/currency.go
